# Boost CLI onLogin pref by proxyClient injection