--- a/lib/client/client.go
+++ b/lib/client/client.go
@@ -752,6 +752,7 @@ func (proxy *ProxyClient) NewWatcher(ctx context.Context, watch types.Watch) (ty
 	return watcher, nil
 }
 
+// GetClusterAlerts connects fetches current clusters alerts using cached current cluster client.
 func (proxy *ProxyClient) GetClusterAlerts(ctx context.Context, req types.GetClusterAlertsRequest) ([]types.ClusterAlert, error) {
 	ctx, span := proxy.Tracer.Start(
 		ctx,