@smallinsky 2023-04-25
@rosstimothy Let me know what do you think about this solution where a proxyClient is initialised in top flow and then propagated by WithProxyClient func so we can fully levrage the [proxy.CurrentCluster()](https://github.com/gravitational/teleport/blob/cf2c7059a3badf64519e90b512d5f6536b1dc848/lib/client/client.go#L1161  ) caching.