@smallinsky 2023-04-26
> What do you think about making the clients explicit arguments in functions that require connections to a cluster instead of using functional options? 

I'm not big fan of this approach without doing complete alignment because we will create a lot of vague functions with nil parameter like `DoSomthing(ctx, tc, nil)`

> I suppose a helper method to get the client from the options might eliminate that.

yea,  I have updated the code applying this approach in: https://github.com/gravitational/teleport/pull/25137/commits/d3c0f6960b592cb76c6719e1cbad99ee1f992274