@smallinsky 2023-05-31
> I took the alternate approach of explicitly requiring an auth.ClientI in https://github.com/gravitational/teleport/pull/27029 - what do you think?

@rosstimothy 
I have thought that this approach will require more alignment but looking at diff changes  it mostly LGTM.  I closing this PR in favour of https://github.com/gravitational/teleport/pull/27029